pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "patch-seg"
authors = [
    { name = "Ian Covert", email = "icovert@stanford.edu" },
    { name = "Tony Sun", email = "tony.sun@cs.stanford.edu" },
]
description = "A probing benchmark for patch-based semantic segmentation"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "timm>=1.0.8",
    "pillow>=10.2.0",
    "numpy>=1.26.4",
    "pandas>=2.2.1",
    "matplotlib>=3.8.3",
    "tqdm?=4.66.2",
]

[project.optional-dependencies]
dev = [
    "black",
    "pre-commit",
    "ruff",
]

[tool.setuptools]
packages = ["patchseg"]

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310"]

[tool.ruff]
line-length = 120
target-version = "py38"
